Move frontend component handling functions to App.js
